userid == 1 can no longer be removed.
git-svn-id: http://svn.cmsmadesimple.org/svn/cmsmadesimple/trunk@2425 3d254a34-79dc-0310-9e5f-be208747d8a0

+++ admin/listusers.php
@@ -41,7 +41,7 @@ if (isset($_GET["message"])) {
 
 	$userid = get_userid();
 	$edit = check_permission($userid, 'Modify Users');
-	$remove = check_permission($userid, 'Remove Users');
+        $remove = check_permission($userid, 'Remove Users');
 
 	$query = "SELECT user_id, username, active FROM ".cms_db_prefix()."users ORDER BY user_id";
 	$result = $db->Execute($query);
@@ -90,7 +90,7 @@ if (isset($_GET["message"])) {
 				    {
 					echo "<td>&nbsp;</td>\n";
 					}
-				if ($remove)
+				if ($remove && $oneuser->id != 1)
 				    {
 					echo "<td><a href=\"deleteuser.php?user_id=".$oneuser->id."\" onclick=\"return confirm('".lang('deleteconfirm')."');\">";
                     echo $themeObject->DisplayImage('icons/system/delete.gif', lang('delete'),'','','systemicon');
